Add switch Amount Exact/Range
Created new state var amountRangeEnabled to control the new switch when advanced options is enabled.

Created memo var makerHasAmountRange to control boolean state of amount exact/range switch using maker.advancedOptions and amountRangeEnabled.

Used makerHasAmountRange instead of maker.advancedOptions when testing if maker is  using amount range or exact amount (changeds places: handleCreateOrder, disableSubmit, amountToString in SummaryText, and in the Amount fields).

Change the default value of amountRageEnabled

Changed the default value of state var "amountRageEnabled" to true, and removed unnecessary use of the method "setAmountRangeEnabled" when chaging between advanced and default options.

Fix style of switch Exact/Range Amount

Fixing the style of the div of the "Exact/Range Amount" Switch removing the unecessary "width" and "height" props and adding a negative marginTop to reduce the gap between buy/sell buttons and the amount input commented in the PR, and added a padding bottom to the label of the amount (exact) to not hide the switch.

Inverting the position of the icon to match the style of the "Advanced Options" Switch.

Fixing the title from tooltip of the switch (the order of the titles was inversed and the text of the Exact Amount was wrong).

Define style of Exact/Range Amount Switch

Last change in the style of the new Exact/Range Amount Switch, following the review recommendations.