Change frontend order detail receiving status message.
Inside the order detail component we have a little text telling the user how much he'sreceiving from the transaction, this commit changes the text removing the lightning word from it because sometimes the receiving method is on-chain. Ideally we would have aorder.is_swap boolean that is only sent to the buyer to identify the method but since we don't have it yet, we are just changing the text.