change current status to ChoiceField
also added ENUM_NAME_OVERRIDES in SPECTACULAR_SETTINGS, because without
it when generating api-latest.yaml this warning is returned:
Warning: encountered multiple names for the same choice set
(CurrentStatusEnum). This may be unwanted even though the generated
schema is technically correct. Add an entry to ENUM_NAME_OVERRIDES to
fix the naming.