add option for timing issues in tests
On my powerful machine, running tests resulted in some errors.
After a bit of digging I found out that it was a timing issues.
This PR adds a config option TIMING_EXTRA_IN_TESTS that when set to
True, it take some time where it should so that the tests do not fail.
If the config option is not present (default) nothing is changed.
I had this issue very sporadically on my laptop, and every time when
running on an other powerful machine.