docs: update Electrum wallet section
Updates the docs section on Electrum wallet. The new version (>= 4.6)
supports anchor channels so it is possible again to open channels to all
lightning nodes.
It now also has the nostr based submarine swap market
which makes it quite suitable for use with RoboSats as it allows to
easily get inbound liquidity.